Changed to listing related actions to MVC
