+++ backend/src/services/listingService.js
@@ -1,0 +1,17 @@
+class ListingService {
+  constructor(prisma) {
+    this.prisma = prisma;
+  }
+
+  async createListing(data) {
+    return this.prisma.rentalListing.create({ data })
+      .catch(e => {
+        console.error(e);
+        return false;
+      });
+  }
+}
+
+module.exports = {
+  ListingService,
+};
